+++ __test__/index.spec.ts
@@ -1,25 +1,14 @@
-import { promises } from 'fs'
+import { promises as fs } from 'fs'
 import { join } from 'path'
 
 import test from 'ava'
-import probeImageSize from 'probe-image-size'
+import jimp from 'jimp'
 
 import { render } from '../index'
 
-test('transparent background', async (t) => {
-  const filePath = './tiger.svg'
-  const svg = await promises.readFile(join(__dirname, filePath))
-  const svgString = svg.toString('utf-8')
-  const pngData = render(svgString)
-  const result = probeImageSize.sync(pngData)
-
-  t.is(result.width, 900)
-  t.is(result.height, 900)
-})
-
 test('fit to width', async (t) => {
   const filePath = '../example/text.svg'
-  const svg = await promises.readFile(join(__dirname, filePath))
+  const svg = await fs.readFile(join(__dirname, filePath))
   const svgString = svg.toString('utf-8')
   const pngData = render(svgString, {
     background: '#eeebe6',
@@ -28,15 +17,39 @@ test('fit to width', async (t) => {
       value: 1200,
     },
   })
-  const result = probeImageSize.sync(pngData)
+  const result = await jimp.read(pngData)
 
-  t.is(result.width, 1200)
-  t.is(result.height, 623)
+  t.is(result.getWidth(), 1200)
+  t.is(result.getHeight(), 623)
+})
+
+test('Set the background with alpha by rgba().', async (t) => {
+  const filePath = './tiger.svg'
+  const svg = await fs.readFile(join(__dirname, filePath))
+  const svgString = svg.toString('utf-8')
+  const pngData = render(svgString, {
+    background: 'rgba(0, 0, 0, 0.6)',
+  })
+  const result = await jimp.read(pngData)
+
+  t.is(result.hasAlpha(), true)
+})
+
+test('Set the background without alpha by hsla()', async (t) => {
+  const filePath = './tiger.svg'
+  const svg = await fs.readFile(join(__dirname, filePath))
+  const svgString = svg.toString('utf-8')
+  const pngData = render(svgString, {
+    background: 'hsla(255, 80%, 50%, 1)',
+  })
+  const result = await jimp.read(pngData)
+
+  t.is(result.hasAlpha(), false)
 })
 
 test('Load custom font', async (t) => {
   const filePath = '../example/text.svg'
-  const svg = await promises.readFile(join(__dirname, filePath))
+  const svg = await fs.readFile(join(__dirname, filePath))
   const svgString = svg.toString('utf-8')
   const pngData = render(svgString, {
     font: {
@@ -45,8 +58,8 @@ test('Load custom font', async (t) => {
       defaultFontFamily: 'Source Han Serif CN Light',
     },
   })
-  const result = probeImageSize.sync(pngData)
+  const result = await jimp.read(pngData)
 
-  t.is(result.width, 1324)
-  t.is(result.height, 687)
+  t.is(result.getWidth(), 1324)
+  t.is(result.getHeight(), 687)
 })